Always pass the defines, SOLUTION by default, to the compiler from check_speedup.py, either 0 or 1
So that the C++ code can determine if a selection has ben done or not.
If a selection has not been made then the C++ code should default to the
optimized version, so that we get proper language analysis of that
preprocessor branch. The suggested

  // Optimized.
  // Baseline.

structure does the opposite, the default, what the IDE sees, is the
baseline code path. Which isn't all that helpful when working on the
optimized code path.