[CompilerIntrinsics4] Update CMakeLists.txt
This fix will help generate identical binaries if no changes were made in the source code. 

Previously, if solution and baseline (in CI) were built in different folders, paths in `data_paths.h` would be different, which would have caused binaries to be different. This change fixes that and identical binaries will be produced.